requirements.txt
mediapipe==0.10.9
opencv-python==4.9.0.80
numpy==1.26.4
scipy==1.12.0
pyyaml==6.0.1
pytest==8.0.0
jupyter==1.0.0
